WEBAPP: show reviewer link only in when status is interview ask for confirmation when updating studies.

+++ metabolights-webapps/src/main/webapp/WEB-INF/jsp/entryActions.jsp
@@ -30,28 +30,28 @@
                         <li><a href="#">change status to ..</a>
                             <ul>
                                 <c:if test="${curator && (study.studyStatus != 'SUBMITTED')}">
-                                <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=SUBMITTED">submitted</a></li>
+                                <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=SUBMITTED" confirmationText="Are you sure you want to change the status back to Submitted?" onclick="return confirmAction(this);">submitted</a></li>
                                 </c:if>
                                 <c:if test="${(study.studyStatus == 'SUBMITTED') || (curator && (study.studyStatus != 'INCURATION'))}">
-                                    <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=INCURATION">in curation</a></li>
+                                    <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=INCURATION" confirmationText="Are you sure you want to send the study to curation?" onclick="return confirmAction(this);">in curation</a></li>
                                 </c:if>
                                 <c:if test="${curator && (study.studyStatus != 'INREVIEW')}">
-                                    <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=INREVIEW">in review</a></li>
+                                    <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=INREVIEW" confirmationText="Is the study ready to be reviewed?" onclick="return confirmAction(this);">in review</a></li>
                                 </c:if>
                                 <c:if test="${curator && (study.studyStatus != 'PUBLIC')}">
-                                    <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=PUBLIC">public</a></li>
+                                    <li><a href="updatestatus?study=${study.studyIdentifier}&newStatus=PUBLIC" confirmationText="This will make the study Publcly available to anyone. Are yo sure?" onclick="return confirmAction(this);">public</a></li>
                                 </c:if>
                             </ul>
                         </li>
                     </c:if>
                     <c:if test="${curator}">
-                        <li><a href="deleteStudy?study=${study.studyIdentifier}">delete</a></li>
+                        <li><a href="deleteStudy?study=${study.studyIdentifier}" confirmationText="This will delete the study from the system, no way back!." onclick="return confirmAction(this);">delete</a></li>
                         <c:if test="${not empty study.backups}">
                             <c:if test="${fn:length(study.backups) gt 0}">
                                 <li><a href="#">restore ...</a>
                                     <ul>
                                         <c:forEach var="backup" items="${study.backups}">
-                                            <li><a href="restore?study=${study.studyIdentifier}&backupidentifier=${backup.backupId}">${backup.backupTimeStamp}</a></li>
+                                            <li><a href="restore?study=${study.studyIdentifier}&backupidentifier=${backup.backupId}" confirmationText="Do you really want to restore the metadata files from a previous version?" onclick="return confirmAction(this);">${backup.backupTimeStamp}</a></li>
                                         </c:forEach>
                                     </ul>
                                 </li>
@@ -63,3 +63,37 @@
         </ul>
     </nav>
 </c:if>
+
+<div id="confirmaction" title="Are you sure..." style="display: none">
+</div>
+<script type="text/javascript">
+   function confirmAction(element){
+
+       var dialog = $("#confirmaction");
+
+       // Fill dialog
+       var targetUrl = $(element).attr("href");
+       var text = $(element).attr("confirmationText");
+
+       dialog.text(text);
+
+       $(dialog).dialog({
+//           autoOpen: false,
+           modal: true,
+           buttons : {
+                   "Confirm" : function() {
+                       window.location.href = targetUrl;
+                   },
+                   "Cancel" : function() {
+                       $(this).dialog("close");
+                   }
+               }
+       });
+
+//       $(dialog).dialog("open");
+
+       return false;
+   }
+
+</script>
+
